+++ src/main/java/org/launchcode/codingevents/models/EventCategory.java
@@ -11,11 +11,8 @@ import java.util.Objects;
  * Created by Chris Bay
  */
 @Entity
-public class EventCategory {
+public class EventCategory extends AbstractEntity {
 
-    @Id
-    @GeneratedValue
-    private int id;
 
     @Size(min=3, message="Name must be at least 3 characters long")
     private String name;
@@ -34,26 +31,11 @@ public class EventCategory {
         this.name = name;
     }
 
-    public int getId() {
-        return id;
-    }
 
     @Override
     public String toString() {
         return name;
     }
 
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        EventCategory that = (EventCategory) o;
-        return id == that.id;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
 }
