+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 # Database connection settings
-spring.datasource.url=jdbc:mysql://localhost:3306/coding_events
+spring.datasource.url=jdbc:mysql://localhost:3306/codingevents
 spring.datasource.username=user
 spring.datasource.password=password
 
@@ -7,7 +7,7 @@ spring.datasource.password=password
 spring.jpa.database = MYSQL
 
 # Show or not log for each sql query
-spring.jpa.show-sql = false
+spring.jpa.show-sql = true
 
 # Hibernate ddl auto (create, create-drop, update)
 spring.jpa.hibernate.ddl-auto = update
